pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forfiles"
version = "4.1.0"
authors = [{ name = "4MBL" }]
description = "Useful tools for files"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["pillow==10.4.0", "layeredimage==2024.3"]
[project.urls]
"Homepage" = "https://github.com/4mbl/forfiles"
